backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
scikit-learn
deep-translator
langdetect
requests
python-dotenv
pydantic